+++ R/load.R
@@ -4,13 +4,12 @@
 #' 
 #' @usage get_datadir(datadir)
 #'   
-#' @param datadir character. Directory as will be passed to 
-#'   \code{\link{load_layers}}. This does not change the data directory used by 
-#'   \code{\link{load_layers}} but only shows the result of passing a specific 
-#'   datadir. If \code{NULL} is used then this returns 
-#'   \code{file.path(path.expand("~"), "R", "sdmpredictors")}. This can be 
-#'   overridden by setting the \code{sdmpredictors_datadir} option with 
-#'   \code{options(sdmpredictors_datadir = "<your preferred directory>")}.
+#' @param datadir character. Directory as passed to \code{\link{load_layers}}.
+#'   This does not change the data directory used by \code{\link{load_layers}}
+#'   but only shows the result of passing a specific datadir. If \code{NULL} is
+#'   passed then the \code{sdmpredictors_datadir} option is read. To set this
+#'   run \code{options(sdmpredictors_datadir = "<your preferred directory>")} in
+#'   every session or in your .RProfile.
 #'   
 #' @return Path to the data directory.
 #' @seealso \code{\link{load_layers}}
@@ -19,15 +18,16 @@ get_datadir <- function(datadir) {
   if(is.null(datadir)) {
     datadir <- getOption("sdmpredictors_datadir")
     if(is.null(datadir)) {
-      datadir <- file.path(path.expand("~"), "R", "sdmpredictors")
-      if(!dir.exists(datadir)) {
-        dir.create(datadir, recursive = TRUE)
-      }
+      datadir <- file.path(tempdir(), "sdmpredictors")
+      warning("file.path(tempdir(), \"sdmpredictors\") will be used as datadir, set options(sdmpredictors_datadir=\"<directory>\") to avoid re-downloading the data in every session or set the datadir parameter in load_layers")
     }
   }
-  normalizePath(paste0(datadir), winslash = "/", mustWork = TRUE)
+  datadir <- normalizePath(paste0(datadir), winslash = "/", mustWork = TRUE)
+  if(!dir.exists(datadir)) {
+    dir.create(datadir, recursive = TRUE)
+  }
+  datadir
 }
-
 
 #' Load layers
 #' 
@@ -35,7 +35,7 @@ get_datadir <- function(datadir) {
 #' RasterStack is returned but this is only possible When all rasters have the 
 #' same spatial extent and resolution.
 #' 
-#' @usage load_layers(layercodes, equalarea=FALSE, rasterstack=TRUE,
+#' @usage load_layers(layercodes, equalarea=FALSE, rasterstack=TRUE, 
 #'   datadir=NULL)
 #'   
 #' @param layercodes character vector or dataframe. Layer_codes of the layers to
@@ -45,18 +45,22 @@ get_datadir <- function(datadir) {
 #'   otherwise unprojected (\code{\link{lonlatproj}}). Default is \code{FALSE}.
 #' @param rasterstack logical. If \code{TRUE} (default value) then the result is
 #'   a \code{\link[raster]{stack}} otherwise a list of rasters is returned.
-#' @param datadir character. Directory where you want to store the data. The 
-#'   default value used is \code{file.path(path.expand("~"), "R", 
-#'   "sdmpredictors")} and can be overridden with 
-#'   \code{options(sdmpredictors_datadir = "<your preferred directory>")}.
+#' @param datadir character. Directory where you want to store the data. If 
+#'   \code{NULL} is passed (default) then the \code{sdmpredictors_datadir} 
+#'   option is read. To set this run \code{options(sdmpredictors_datadir="<your 
+#'   preferred directory>")} in every session or add it to your .RProfile.
 #'   
 #' @return RasterStack or list of raster
-#'   
+#' @examples \dontrun{
+#' # warning using tempdir() implies that data will be downloaded again in the 
+#' # next R session
+#' env <- load_layers("BO_calcite", datadir = tempdir())
+#' }
 #' @export
 #' @seealso \code{\link{list_layers}}, \code{\link{layer_stats}}, 
 #'   \code{\link{layers_correlation}}
 load_layers <- function(layercodes, equalarea = FALSE, rasterstack = TRUE, datadir = NULL) {
-  if(is.na(equalarea) || !is.logical(equalarea) && length(equalarea) != 1) {
+  if(is.na(equalarea) || !is.logical(equalarea) || length(equalarea) != 1) {
     stop("equalarea should be TRUE or FALSE")
   }
   if(is.data.frame(layercodes)) {
@@ -80,9 +84,15 @@ load_layers <- function(layercodes, equalarea = FALSE, rasterstack = TRUE, datad
     suffix <- ifelse(equalarea, "", "_lonlat")
     path <- paste0(datadir, "/", layercode, suffix, ".tif")
     if(!file.exists(path)) {
-      urlroot <- "http://www.phycology.ugent.be/research/sdmpredictors/"
       url <- paste0(urlroot, layercode, suffix, ".tif")
-      download.file(url, path, method = "auto", quiet = FALSE, mode = "wb")
+      ok <- -1
+      # clean up of download failed
+      on.exit({
+        if(ok != 0 && file.exists(path)) {
+          file.remove(path)
+        }
+      })
+      ok <- utils::download.file(url, path, method = "auto", quiet = FALSE, mode = "wb")  
     }
     ifelse(file.exists(path), path, NA)
   }