+++ tests/testthat/test_get_sysdata.R
@@ -2,6 +2,8 @@ library(sdmpredictors)
 
 context("Get sysdata")
 
+
+
 test_that("get_sysdata has no unnamed list entries", {
   sysdata <- sdmpredictors:::get_sysdata()
   expect_false(any(names(sysdata) == ""))
@@ -20,16 +22,24 @@ test_that("get_sysdata returns all expected entries", {
   expect_notnull(sysdata$layerscorrelation)
   expect_notnull(sysdata$urldata)
   expect_notnull(sysdata$urlsysdata)
-  expect_equal(length(names(sysdata)), 8)
+  expect_notnull(sysdata$bibentries)
+  expect_equal(length(names(sysdata)), 9)
 })
 
 test_that("sysdata gets downloaded", {
-  fname <- "sysdata.rda"
-  outfile <- paste0(sdmpredictors:::get_datadir(NULL), "/", fname)
-  if(file.exists(outfile)) {
-    file.remove(outfile)
-    expect_false(file.exists(outfile))
+  gets_downloaded <- function(outfile) {
+    if(file.exists(outfile)) {
+      file.remove(outfile)
+      expect_false(file.exists(outfile))
+    }
+    sysdata <- sdmpredictors:::get_sysdata()
+    expect_true(file.exists(outfile))
   }
-  sysdata <- sdmpredictors:::get_sysdata()
-  expect_true(file.exists(outfile))
+  fname <- "sysdata.rda"
+  options(sdmpredictors_datadir = file.path(tempdir(), "sdmpredictors_test"))
+  outfile <- paste0(getOption("sdmpredictors_datadir"), "/", fname)
+  gets_downloaded(outfile)
+  options(sdmpredictors_datadir = NULL)
+  outfile <- paste0(tempdir(), "/sdmpredictors/", fname)
+  gets_downloaded(outfile)
 })