+++ NEWS.md
@@ -1,3 +1,28 @@
+# sdmpredictors 0.2.5
+
+Fix authors
+
+# sdmpredictors 0.2.4
+
+New datasets (ENVIREM, WorldClim paleo and future)
+Added functions related to correlations
+
+# sdmpredictors 0.2.3
+
+Remove usage of ~/R/sdmpredictors from tests
+
+# sdmpredictors 0.2.2
+
+Fix url in description
+
+# sdmpredictors 0.2.1
+
+Fix urls in description and readme
+
+# sdmpredictors 0.2
+
+Datadir is mandatory now, instead of automatically writing to ~/R/sdmpredictors.
+
 # sdmpredictors 0.1
 
 Initial release of the sdmpredictors package.